monitor/requirements.txt
# YouTube Data API v3
google-api-python-client>=2.100.0
google-auth>=2.23.0

# Config
PyYAML>=6.0.1

# TikTok：HTTP 快路 + Playwright（TikTokApi）
httpx>=0.27.0
yt-dlp>=2024.12.0
TikTokApi>=7.3.0
playwright>=1.40.0

# Web 看板
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
